Issue #45
Adding functionality to address the saved instances when the
.hab_user_prefs.json timestamp expires.  Currently Hab will just error
out.

I've addessed this in both user_prefs.py and cli.py.  In user_prefs.py
I've replaced the "uri_reason" method with a method named "uri_check".
This method passes data onto the cli.py in the form of a namedtuple.
I then moved the job of naming the "reason" over into the cli.py.

In the cli.py, I then am taking the data from uri_check to determine
if the user_prefs timestamp has expire, then I provide the user with the
option to resave the prefs with an updated timestamp then continue using
Hab.  If the user declines to resave, Hab will quit with a
click.UsageError message.