Add pid to batch's random filename
Batch's `%RANDOM%` resolution is limited to about a second, so if you are
calling hab multiple times at once you will end up with more than one hab
process using the same temp files and causing errors when the first one
to finish cleans up after itself.

Add the pid to the dirname to address this.