hab_base._collect_base now fully processes each property in order
Switch from processing all properties on the current node, then checking
the parent if any were NotSet to fully processing each property in
sort_key order before moving on to the next property.

This fixes an bug where distro global env vars are not added to the global env when distro is inherited.
This also makes it so hab_properties can be defined only on parent or default.