1/5: Add DistroFinder class
This class will replace the current code used to find distros. This allows
us to implement sub-classes to change how they are found. For example to
allow us to resolve dependencies hosted on the cloud for download and install.

Add `hab.utils.glob_path` that supports globbing a pathlib.Path object
with glob strings in the path.