1/2: Prevent hab env from exiting if any errors happen in the child shell
The config.sh template now keeps track of if it needed to call `set -e`
and ensures if so it is reverted before exiting the script even on error.

I'm unable to add automated testing so I've created some manual test scripts
that guide testing of this feature. These should be run periodically.

The problem seems to be that when using subprocess run a hab launch command
in bash doesn't work. All of the bash code works and its able to use hab
to generate the .sh files, but it silently doesn't call the launched alias
and doesn't error out either.